Fetch openapi spec dynamically
