+++ .gitignore
@@ -33,6 +33,7 @@ packages/state-transition/test-cache
 benchmark_data
 invalidSszObjects/
 packages/beacon-node/mainnet_pubkeys.csv
+packages/api/oapi-schemas
 
 # Autogenerated docs
 packages/**/docs
