+++ packages/api/test/perf/compileRouteUrlFormater.test.ts
@@ -1,4 +1,4 @@
-import {compileRouteUrlFormater} from "../src/utils/urlFormat.js";
+import {compileRouteUrlFormater} from "../../src/utils/urlFormat.js";
 
 /* eslint-disable no-console */
